Remove separate JvmTestSuitePluginHelper
Define default test suite name in the test suites plugin itself.  Move method back to JavaPluginHelper.
