+++ subprojects/plugins/src/main/java/org/gradle/api/plugins/JavaLibraryPlugin.java
@@ -18,7 +18,6 @@ package org.gradle.api.plugins;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
 import org.gradle.api.plugins.internal.JavaPluginHelper;
-import org.gradle.api.plugins.internal.JvmTestSuitePluginHelper;
 import org.gradle.api.plugins.jvm.JvmTestSuite;
 import org.gradle.jvm.component.internal.JvmSoftwareComponentInternal;
 
@@ -44,7 +43,7 @@ public abstract class JavaLibraryPlugin implements Plugin<Project> {
         component.getMainFeature().withApi();
 
         // Make compileOnlyApi visible to tests.
-        JvmTestSuite defaultTestSuite = JvmTestSuitePluginHelper.getDefaultTestSuite(project);
+        JvmTestSuite defaultTestSuite = JavaPluginHelper.getDefaultTestSuite(project);
         project.getConfigurations()
             .getByName(defaultTestSuite.getSources().getCompileOnlyConfigurationName())
             .extendsFrom(component.getMainFeature().getCompileOnlyApiConfiguration());
